+++ src/crepes/__init__.py
@@ -1,15 +1,2 @@
-"""Conformal regressors and predictive systems (crepes)
-
-Classes implementing conformal regressors and conformal predictive
-systems, which transform point predictions into prediction intervals
-and cumulative distribution functions, respectively.
-
-Author: Henrik Boström (bostromh@kth.se)
-
-Copyright 2022 Henrik Boström
-
-License: BSD 3 clause
-"""
-from crepes.base import ConformalRegressor, ConformalPredictiveSystem, __version__
-
+from crepes.base import ConformalRegressor, ConformalPredictiveSystem, ConformalPredictor, __version__